Disable checks in bsg_manycore_features.h
I don't think any of these checks are necessary for emulation. (I am not
even sure they are necessary for the proper manycore libraries!) And
including <features.h> is a problem on macOS, which doesn't have it.
(That header is a glibc-specific thing, not a POSIX C thing.)

This is part of #40.